+++ examples/callback_awaitable.hpp
@@ -202,7 +202,7 @@ template <IsAwCallback Awaitable> struct awaitable_traits<Awaitable> {
     awaitable.customizer.done_count = DoneCount;
   }
 
-  static void set_flags(self_type& awaitable, uint64_t Flags) {
+  static void set_flags(self_type& awaitable, size_t Flags) {
     awaitable.customizer.flags = Flags;
   }
 };
