+++ examples/spawn_misc.cpp
@@ -18,13 +18,13 @@ template <size_t Count, size_t ThreadCount> void small_func_spawn_bench_lazy() {
   std::printf("small_func_spawn_bench_lazy()...\n");
   ex_cpu executor;
   executor.set_thread_count(ThreadCount).init();
-  std::array<uint64_t, Count> data;
+  std::array<size_t, Count> data;
   for (size_t i = 0; i < Count; ++i) {
     data[i] = 0;
   }
   std::array<std::future<void>, Count> results;
   auto preTime = std::chrono::high_resolution_clock::now();
-  for (uint64_t i = 0; i < Count; ++i) {
+  for (size_t i = 0; i < Count; ++i) {
     // because this is a functor and not a coroutine,
     // it is OK to capture the loop variables
     results[i] = post_waitable(executor, [i, &data]() { data[i] = i; }, 0);
@@ -35,7 +35,7 @@ template <size_t Count, size_t ThreadCount> void small_func_spawn_bench_lazy() {
   }
   auto doneTime = std::chrono::high_resolution_clock::now();
 
-  for (uint64_t i = 0; i < Count; ++i) {
+  for (size_t i = 0; i < Count; ++i) {
     if (data[i] != i) {
       std::printf("FAIL: index %" PRIu64 " value %" PRIu64 "", i, data[i]);
     }
@@ -63,19 +63,19 @@ template <size_t Count, size_t nthreads> void large_task_spawn_bench_lazy() {
   std::printf("large_task_spawn_bench_lazy()...\n");
   ex_cpu executor;
   executor.set_thread_count(nthreads).init();
-  std::array<uint64_t, Count> data;
+  std::array<size_t, Count> data;
   for (size_t i = 0; i < Count; ++i) {
     data[i] = 0;
   }
   std::array<std::future<void>, Count> results;
   auto preTime = std::chrono::high_resolution_clock::now();
-  for (uint64_t slot = 0; slot < Count; ++slot) {
+  for (size_t slot = 0; slot < Count; ++slot) {
     // because this is a coroutine and not a functor, it is not safe to capture
     // https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/avoid-capturing-lambda-coroutines.html
     // variables must be passed as parameters instead
     results[slot] = post_waitable(
       executor,
-      [](uint64_t* DataSlot) -> task<void> {
+      [](size_t* DataSlot) -> task<void> {
         int a = 0;
         int b = 1;
         for (int i = 0; i < 1000; ++i) {
@@ -119,13 +119,13 @@ void large_task_spawn_bench_lazy_bulk() {
   std::printf("large_task_spawn_bench_lazy_bulk()...\n");
   ex_cpu executor;
   executor.set_thread_count(nthreads).init();
-  std::array<uint64_t, Count> data;
+  std::array<size_t, Count> data;
   for (size_t i = 0; i < Count; ++i) {
     data[i] = 0;
   }
   auto preTime = std::chrono::high_resolution_clock::now();
   auto tasks =
-    std::ranges::views::transform(data, [](uint64_t& DataSlot) -> task<void> {
+    std::ranges::views::transform(data, [](size_t& DataSlot) -> task<void> {
       int a = 0;
       int b = 1;
       for (int i = 0; i < 1000; ++i) {
@@ -166,7 +166,7 @@ void prio_reversal_test() {
   std::printf("prio_reversal_test()...\n");
   ex_cpu executor;
   executor.set_thread_count(nthreads).set_priority_count(npriorities).init();
-  std::array<uint64_t, Count> data;
+  std::array<size_t, Count> data;
   for (size_t i = 0; i < Count; ++i) {
     data[i] = 0;
   }
@@ -174,10 +174,10 @@ void prio_reversal_test() {
   auto preTime = std::chrono::high_resolution_clock::now();
   size_t slot = 0;
   while (true) {
-    for (uint64_t prio = npriorities - 1; prio != -1ULL; --prio) {
+    for (size_t prio = npriorities - 1; prio != -1ULL; --prio) {
       results[slot] = post_waitable(
         executor,
-        [](uint64_t* DataSlot, [[maybe_unused]] size_t Priority) -> task<void> {
+        [](size_t* DataSlot, [[maybe_unused]] size_t Priority) -> task<void> {
           int a = 0;
           int b = 1;
           for (int i = 0; i < 1000; ++i) {
